# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "cssthema API",
    "description": "Central CSS Theme Builder — zie docs/05-api-specificatie.md",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/v1/themes": {
      "get": {
        "tags": [
          "themes"
        ],
        "summary": "List Themes",
        "operationId": "themes_list",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "description": "Zoekt in naam, slug en beschrijving.",
              "title": "Q"
            },
            "description": "Zoekt in naam, slug en beschrijving."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ThemeStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "palette_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Palette Id"
            }
          },
          {
            "name": "service_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Service Id"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 40
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "name": "include_deleted",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Deleted"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "-updated_at",
                "name",
                "-created_at"
              ],
              "type": "string",
              "default": "-updated_at",
              "title": "Sort"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 1024
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThemePage"
                }
              }
            }
          },
          "500": {
            "description": "Interne fout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "themes"
        ],
        "summary": "Create Theme",
        "operationId": "themes_create",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThemeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Theme"
                }
              }
            }
          },
          "500": {
            "description": "Interne fout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Conflict (slug of status)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "CSS groter dan CSS_MAX_BYTES",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Ongeldig",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/themes/import": {
      "post": {
        "tags": [
          "themes"
        ],
        "summary": "Import Theme",
        "operationId": "themes_import",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_themes_import"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Theme"
                }
              }
            }
          },
          "500": {
            "description": "Interne fout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "200": {
            "description": "Nieuwe versie op een bestaand thema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Theme"
                }
              }
            }
          },
          "409": {
            "description": "Conflict (slug of status)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "CSS groter dan CSS_MAX_BYTES",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "415": {
            "description": "Geen .css of .cssthema.zip",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Lint-fouten of ongeldige invoer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LintFailedProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/themes/local-files": {
      "get": {
        "tags": [
          "themes"
        ],
        "summary": "Handgemaakte CSS-bestanden in CSS_FILES_DIR",
        "operationId": "themes_list_local_files",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LocalCssFile"
                  },
                  "type": "array",
                  "title": "Response Themes List Local Files"
                }
              }
            }
          },
          "500": {
            "description": "Interne fout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/themes/local-files/import": {
      "post": {
        "tags": [
          "themes"
        ],
        "summary": "Handgemaakte CSS-bestanden als thema importeren (en archiveren)",
        "operationId": "themes_import_local_files",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalImportResult"
                }
              }
            }
          },
          "500": {
            "description": "Interne fout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/themes/{theme_id}/restore": {
      "post": {
        "tags": [
          "themes"
        ],
        "summary": "Restore Theme",
        "operationId": "themes_restore",
        "parameters": [
          {
            "name": "theme_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Theme Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Theme"
                }
              }
            }
          },
          "500": {
            "description": "Interne fout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Thema niet gevonden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Slug intussen bezet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/themes/{theme_id}/duplicate": {
      "post": {
        "tags": [
          "themes"
        ],
        "summary": "Duplicate Theme",
        "operationId": "themes_duplicate",
        "parameters": [
          {
            "name": "theme_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Theme Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DuplicateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Theme"
                }
              }
            }
          },
          "500": {
            "description": "Interne fout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Thema niet gevonden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Conflict (slug of status)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/themes/{theme_id}/versions/{version_number}": {
      "get": {
        "tags": [
          "themes"
        ],
        "summary": "Get Version",
        "operationId": "themes_get_version",
        "parameters": [
          {
            "name": "theme_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Theme Id"
            }
          },
          {
            "name": "version_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2147483647,
              "minimum": 1,
              "title": "Version Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Version"
                }
              }
            }
          },
          "500": {
            "description": "Interne fout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Thema niet gevonden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LocalImportRequest": {
        "properties": {
          "names": {
            "items": {
              "type": "string",
              "maxLength": 255
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Names"
          },
          "publish": {
            "type": "boolean",
            "title": "Publish",
            "default": true
          },
          "archive": {
            "type": "boolean",
            "title": "Archive",
            "description": "Verplaats het bestand na de import naar `.geimporteerd/`, zodat nginx naar het thema doorvalt. Alleen als het thema live is (`publish`); anders blijft het bestand staan en staat de reden in `archive_error`.",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "names"
        ],
        "title": "LocalImportRequest"
      },
...
      "ThemePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Theme"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ThemePage"
      },
      "ThemeStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published",
          "archived"
        ],
        "title": "ThemeStatus"
      },
...
    }
  }
}